10,000 Population, P(Frag. = 0.20)
CM = Constant coalescence, multiplicative fragmentation. CC = Constant coalescence, constant fragmentation.
ImpFrag is imperfect fragmentation, with many varieties. BB is a beta-binomial with coefficients X and Y. Total X of Y Frag is to shatter X/Y of a cluster.
ImpCo and ImpCoFr refer to imperfect fragmentation with BB 1 and 1, but with the fragmentation alternatively or also affecting the coalescence process.